+++ packages/webpack5/src/configuration/angular.ts
@@ -1,8 +1,9 @@
 import Config from 'webpack-chain';
 import path from 'path';
 
-import { getEntryPath, getProjectRootPath } from '../helpers/project';
+import { getProjectRootPath } from '../helpers/project';
 import { env as _env, IWebpackEnv } from '../index';
+import { getEntryPath } from '../helpers/platform';
 import base from './base';
 
 export default function (config: Config, env: IWebpackEnv = _env): Config {
