+++ packages/webpack5/src/configuration/vue.ts
@@ -3,7 +3,7 @@ import { merge } from 'webpack-merge';
 import Config from 'webpack-chain';
 
 import { env as _env, IWebpackEnv } from '../index';
-import { getPlatformName } from '../platforms';
+import { getPlatformName } from '../helpers/platform';
 import base from './base';
 
 export default function (config: Config, env: IWebpackEnv = _env): Config {
