+++ packages/webpack5/src/helpers/log.ts
@@ -15,7 +15,9 @@ export function error(...data: any): Error {
 
 	// we return the error - the caller can throw or ignore
 	if (typeof data[0] === 'string') {
-		return new Error(data[0]);
+		return new Error(
+			'\n\n[@nativescript/webpack]\n---\n\n' + dedent(data[0]) + '\n\n---\n'
+		);
 	}
 
 	return new Error('@nativescript/webpack ran into a problem...');
