+++ packages/webpack5/src/helpers/project.ts
@@ -1,43 +1,7 @@
-import { resolve, dirname } from 'path';
-
-import { getPlatform } from '../platforms';
+import { resolve } from 'path';
 
 export function getProjectRootPath(): string {
 	return process.cwd();
-}
-
-export function getEntryPath() {
-	const platform = getPlatform();
-
-	// use platform specific entry path
-	if (platform.getEntryPath) {
-		return platform.getEntryPath();
-	}
-
-	// fallback to main field in package.json
-	const packageJson = getPackageJson();
-
-	return resolve(getProjectRootPath(), packageJson.main);
-}
-
-export function getEntryDirPath() {
-	return dirname(getEntryPath());
-}
-
-export function getDistPath() {
-	const platform = getPlatform();
-
-	// use platform specific entry path
-	if (platform.getDistPath) {
-		return platform.getDistPath();
-	}
-
-	// fallback to a generic dist folder
-	return 'dist';
-}
-
-export function getAbsoluteDistPath() {
-	return resolve(getProjectRootPath(), getDistPath());
 }
 
 interface IPackageJson {
